Time: 3 ms (95.14%) | Memory: 10.4 MB (77.96%) - LeetSync

+++ 518-coin-change-ii/coin-change-ii.cpp
@@ -4,16 +4,16 @@ public:
     int change(int amount, vector<int>& coins) {
         int n = coins.size();
 
-        vector<unsigned long long> dp(amount + 1, 0);
+        vector<double> dp(amount + 1, 0);
 
         dp[0] = 1;
 
-        for(long long coin : coins) {
-            for(long long t = coin; t <= amount; t++) {
+        for(int coin : coins) {
+            for(int t = coin; t <= amount; t++) {
                 dp[t] = dp[t] + dp[t - coin];
             }
         }
 
-        return dp[amount];
+        return (int)dp[amount];
     }
 };
